chore: move gift card data updates to transactionMetadataProvider
